Home: split solution constants and types
