+++ public/app/features/home/solutions/constants.ts
@@ -1,0 +1,1 @@
+export const SOLUTION_IDS = ['kubernetes', 'metrics', 'logs', 'traces'] as const;
